requirements.txt
pymongo[srv]==3.10
python-dotenv